% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{methods}
\name{getLocus}
\alias{getAccession}
\alias{getComment}
\alias{getDBLink}
\alias{getDBSource}
\alias{getDate}
\alias{getDefinition}
\alias{getDivision}
\alias{getGeneID}
\alias{getKeywords}
\alias{getLength}
\alias{getLocus}
\alias{getMoltype}
\alias{getOrganism}
\alias{getReference}
\alias{getSource}
\alias{getTaxonomy}
\alias{getTopology}
\alias{getVersion}
\title{Access the various fields of a GenBank record.}
\usage{
getLocus(x)

getLength(x)

getMoltype(x)

getTopology(x)

getDivision(x)

getDate(x)

getDefinition(x)

getAccession(x)

getVersion(x)

getGeneID(x)

getDBLink(x)

getDBSource(x)

getSource(x)

getOrganism(x)

getTaxonomy(x)

getKeywords(x)

getReference(x)

getComment(x)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}},
  \code{\linkS4class{gbFeatureTable}},
  \code{\linkS4class{gbRecord}}, or
  \code{\linkS4class{gbRecordList}} object.}
}
\description{
Access the various fields of a GenBank record.
}
\examples{
gbk_file <- system.file("extdata", "marine_metagenome.gbk", package = "biofiles")
x <- gbRecord(gbk_file)

getLocus(x)
getLength(x)
getGeneID(x)
getReference(x)
}

