\docType{methods}
\name{getLocus}
\alias{getAccession}
\alias{getComment}
\alias{getDBLink}
\alias{getDBSource}
\alias{getDate}
\alias{getDefinition}
\alias{getDivision}
\alias{getGeneID}
\alias{getKeywords}
\alias{getLength}
\alias{getLocus}
\alias{getMoltype}
\alias{getOrganism}
\alias{getReference}
\alias{getSource}
\alias{getTaxonomy}
\alias{getTopology}
\alias{getVersion}
\title{Access slots of a GenBank record.}
\usage{
getLocus(x)

getLength(x)

getMoltype(x)

getTopology(x)

getDivision(x)

getDate(x)

getDefinition(x)

getAccession(x)

getVersion(x)

getGeneID(x)

getDBLink(x)

getDBSource(x)

getSource(x)

getOrganism(x)

getTaxonomy(x)

getKeywords(x)

getReference(x)

getComment(x)
}
\description{
Access slots of a GenBank record.
}

