% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbLocation-class.R
\name{as.gbLocation}
\alias{as.gbLocation}
\title{Create a \code{gbLocation}.}
\usage{
as.gbLocation(base_span)
}
\arguments{
\item{base_span}{A character string representation of GenBank feature location}
}
\value{
A \code{\linkS4class{gbLocation}} object.
}
\description{
Create a \code{gbLocation} object out of a character string.
}
\examples{
as.gbLocation("join(1..10,12..20)")
}

