% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{package}
\name{biofiles-package}
\alias{biofiles-package}
\title{This is an R package for interfacing with GenBank records}
\description{
This is an R package for interfacing with GenBank and
GenPept flat file records. It includes utilities for
reading and writing GenBank files, and methods for
interacting with annotation and sequence data.
}
\author{
Gerhard Schöfl \email{gschofl@yahoo.de}
}
\references{
For a sample GenBank record see
\url{http://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html}

For a detailed description of the GenBank feature table
format see \url{http://www.ncbi.nlm.nih.gov/collab/FT}
}
\keyword{package}

