% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFiles-package.R
\docType{package}
\name{BioFiles}
\alias{BioFiles}
\alias{BioFiles-package}
\title{BioFiles - Work with bioinformatic files}
\description{
Read, write, search, modify, and convert between various bioinformatic files.
Includes objects for files and entries, as well as validators. [SRJ:: At
least it will eventually]
}

