% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{.filter}
\alias{.filter}
\title{Filter gbFeatures from a FeatureTable}
\usage{
.filter(x, ..., .cols = NULL)
}
\arguments{
\item{x}{A gbFeatureTable}

\item{...}{key=value pairs interpreted as filters}

\item{.cols}{gbFeature annotations returned as a data.frame}
}
\description{
Filter gbFeatures from a FeatureTable
}
\keyword{internal}

