% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{.select}
\alias{.select}
\title{Select annotations from a FeatureTable}
\usage{
.select(x, ..., .cols = NULL)
}
\arguments{
\item{x}{A gbFeatureTable}

\item{...}{character keys}

\item{.cols}{gbFeature annotations returned as a data.frame}
}
\description{
Select annotations from a FeatureTable
}
\keyword{internal}

