% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{embl_record}
\alias{embl_record}
\title{Parser for Embl Records.}
\usage{
embl_record(rec)
}
\arguments{
\item{x}{A character vector}
}
\value{
A \code{\linkS4class{gbRecord}} instance.
}
\description{
Parser for Embl Records.
}
\keyword{internal}

