% Built by roxygen3 0.1
\docType{methods}
\name{end,gbFeature-method}
\alias{end,gbFeature-method}
\title{Get the end of genomic features}
\usage{
  \S4method{end}{gbFeature}(x, ...)
}
\arguments{
  \item{x}{A \code{gbFeature} object.}

  \item{join}{Join compound genomic locations into a single range.}

  \item{...}{Further arguments passed to methods.}
}

\value{
An integer vector
}
\description{
Get the end of genomic features
}
\keyword{methods}
