% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeature-class.R, R/gbFeatureTable-class.R, R/gbLocation-class.R, R/gbRecord-class.R, R/gbRecordList-class.R
\docType{methods}
\name{end}
\alias{end}
\alias{end,gbFeature-method}
\alias{end,gbFeatureTable-method}
\alias{end,gbLocation-method}
\alias{end,gbRecord-method}
\alias{end,gbRecordList-method}
\alias{end<-}
\alias{end<-,gbFeature-method}
\alias{end<-,gbFeatureTable-method}
\alias{end<-,gbLocation-method}
\title{Get or set the end position of genomic features}
\usage{
end(x, ...)

end(x, ...) <- value

\S4method{end}{gbFeature}(x, join = FALSE)

\S4method{end}{gbFeature}(x, ...) <- value

\S4method{end}{gbFeatureTable}(x, join = FALSE)

\S4method{end}{gbFeatureTable}(x, ...) <- value

\S4method{end}{gbLocation}(x, join = FALSE)

\S4method{end}{gbLocation}(x, ...) <- value

\S4method{end}{gbRecord}(x, join = FALSE)

\S4method{end}{gbRecordList}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{value}{The end information to set on \code{x}.}

\item{join}{Join compound genomic locations into a single range.}

\item{check}{if \code{FALSE}, don't perform validity checks.}
}
\value{
An integer vector or a list of integer vectors.
}
\description{
Get or set the end position of genomic features
}
\section{Methods (by class)}{
\itemize{
\item \code{gbFeature}: 

\item \code{gbFeature}: 

\item \code{gbFeatureTable}: 

\item \code{gbFeatureTable}: 

\item \code{gbLocation}: 

\item \code{gbLocation}: 

\item \code{gbRecord}: 

\item \code{gbRecordList}: 
}}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))

end(x)
cds <- x["CDS"]
end(cds)
}
\seealso{
\code{\link{start}}, \code{\link{strand}}, \code{\link{span}}, \code{\link{ranges}}
}

