\docType{methods}
\name{getFeatures}
\alias{ft}
\alias{getFeatures}
\title{Get the feature table from a GenBank record.}
\usage{
getFeatures(x, ...)

ft(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
The \code{\linkS4class{gbFeatureList}} of a Genbank record.
}
\description{
Get the feature table from a GenBank record.
}

