\name{features}
\alias{features}
\alias{features-methods}
\title{Retrieve features of a GenBank record.}
\usage{
  features(x, ...)
}
\arguments{
  \item{data}{An instance of
  \code{\linkS4class{gbRecord}}.}
}
\value{
  The \code{\linkS4class{gbFeatureList}} object contained
  in a gbRecord database.
}
\description{
  Retrieve features of a GenBank record.
}

