% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{methods}
\name{fuzzy}
\alias{fuzzy}
\alias{fuzzy,gbFeature-method}
\alias{fuzzy,gbFeatureTable-method}
\alias{fuzzy,gbRecord-method}
\alias{fuzzy,gbRecordList-method}
\title{Has a feature unclear start/end positions?}
\usage{
fuzzy(x)

\S4method{fuzzy}{gbFeature}(x)

\S4method{fuzzy}{gbFeatureTable}(x)

\S4method{fuzzy}{gbRecord}(x)

\S4method{fuzzy}{gbRecordList}(x)
}
\description{
Has a feature unclear start/end positions?
}

