% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-gbk.R
\name{gbk_record}
\alias{gbk_record}
\title{Parser for GenBank/GenPept records.}
\usage{
gbk_record(rec)
}
\arguments{
\item{x}{A character vector}
}
\value{
A \code{\linkS4class{gbRecord}} instance.
}
\description{
Parser for GenBank/GenPept records.
}
\keyword{internal}

