% Built by roxygen3 0.1
\docType{class}
\name{gbRange-class}
\alias{gbRange-class}
\title{gbRange class}
\description{
gbRange class
}
\section{Class hierarchy for gbRange}{
Super classes:
\itemize{
  \item \linkS4class{IRanges}
  \item \linkS4class{Ranges}
  \item \linkS4class{IntegerList}
  \item \linkS4class{RangesORmissing}
  \item \linkS4class{AtomicList}
  \item \linkS4class{List}
  \item \linkS4class{Vector}
  \item \linkS4class{Annotated}
}

}

\section{Generics with methods for gbRange}{
\itemize{
  \item \code{\link{[}}
  \item \code{\link{[[}}
  \item \code{\link{$}}
  \item \code{\link{annotation}}
  \item \code{\link{end}}
  \item \code{\link{initialize}}
  \item \code{\link{range}}
  \item \code{\link{show}}
  \item \code{\link{start}}
}

}

\keyword{classes}
