\docType{methods}
\name{getFeatures}
\alias{getFeatures}
\title{Retrieve features of a GenBank record.}
\arguments{
  \item{data}{An instance of \code{\link{gbRecord-class}}.}
}
\value{
  The \code{gbFeatureList-class} object contained in a
  gbRecord database.
}
\description{
  Retrieve features of a GenBank record.
}
\seealso{
  Other accessors: \code{\link{gbFeature-class}},
  \code{\link{getIndex}}, \code{\link{getKey}},
  \code{\link{getLocation}}, \code{\link{getQualifier}},
  \code{\link{getSequence}}, \code{\link{hasKey}},
  \code{\link{hasQualifier}}
}

