\docType{class}
\name{gbRecordList-class}
\alias{gbRecordList}
\alias{gbRecordList-class}
\title{gbRecordList-class}
\usage{
gbRecordList(...)
}
\arguments{
  \item{...}{\dQuote{\linkS4class{gbRecord}} elements.}
}
\description{
\dQuote{gbRecordList} is an S4 class that provides a
container for \dQuote{\linkS4class{gbRecord}}s retrived
from GenBank flat files.
}

