% Built by roxygen3 0.1
\docType{class}
\name{gbRecordList-class}
\alias{gbRecordList}
\alias{gbRecordList-class}
\title{gbRecordList}
\usage{
  gbRecordList(...)
}
\arguments{
  \item{...}{\dQuote{\linkS4class{gbRecord}} elements.}
}

\description{
\dQuote{gbRecordList} is an S4 class that provides a container for
\dQuote{\linkS4class{gbRecord}}s retrived from GenBank flat files.
}
\section{Class hierarchy for gbRecordList}{
Super classes:
\itemize{
  \item \linkS4class{list}
  \item \linkS4class{vector}
}

}

\section{Generics with methods for gbRecordList}{
\itemize{
  \item \code{\link{.dbSource}}
  \item \code{\link{.defline}}
  \item \code{\link{end}}
  \item \code{\link{getAccession}}
  \item \code{\link{getComment}}
  \item \code{\link{getDate}}
  \item \code{\link{getDBLink}}
  \item \code{\link{getDBSource}}
  \item \code{\link{getDefinition}}
  \item \code{\link{getDivision}}
  \item \code{\link{getFeatures}}
  \item \code{\link{getGeneID}}
  \item \code{\link{getKeywords}}
  \item \code{\link{getLength}}
  \item \code{\link{getLocus}}
  \item \code{\link{getMoltype}}
  \item \code{\link{getOrganism}}
  \item \code{\link{getReference}}
  \item \code{\link{getSequence}}
  \item \code{\link{getSource}}
  \item \code{\link{getTaxonomy}}
  \item \code{\link{getTopology}}
  \item \code{\link{getVersion}}
  \item \code{\link{ranges}}
  \item \code{\link{show}}
  \item \code{\link{start}}
  \item \code{\link{strand}}
  \item \code{\link{summary}}
  \item \code{\link{width}}
}

}

\keyword{classes}
