\name{gbReferenceList-class}
\alias{gbReferenceList-class}
\title{Class \code{"gbReferenceList"}}
\description{
A container for a set of GenBank REFERENCE fields.
}
\section{Fields}{
  \describe{ \item{\code{ref}:}{ A list of
  \code{\linkS4class{gbReference}} objects. } }
}

\section{Extends}{
  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("gbReferenceList")
}
\seealso{
\code{\link{.gbReferenceList}}
}
\keyword{classes}

