\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures-methods}
\title{Retrieve features of a GenBank record.}
\usage{
  getFeatures(x, ...)
}
\arguments{
  \item{data}{An instance of \code{\link{gbRecord-class}}.}
}
\value{
  The \code{\link{gbFeatureList-class}} object contained in
  a gbRecord database.
}
\description{
  Retrieve features of a GenBank record.
}

