\name{getIndex}
\alias{getIndex}
\alias{getIndex-methods}
\title{Retrieve feature indices from a GenBank record}
\usage{
  getIndex(x, attributes = FALSE)
}
\arguments{
  \item{x}{A \code{gbFeature} or \code{gbFeatureList}
  object.}

  \item{attributes}{Set the \code{accession},
  \code{definition}, \code{database} attributes.}
}
\value{
  A numeric vector of feature indeces
}
\description{
  Retrieve feature indices from a GenBank record
}

