\name{hasValidDb}
\alias{hasValidDb}
\title{validate a gbRecord database which if referenced by gbFeature or
gbFeatureList objects by their '.Dir' slot.}
\usage{
  hasValidDb(x, verbose = TRUE, fields = GBFIELDS)
}
\description{
  validate a gbRecord database which if referenced by
  gbFeature or gbFeatureList objects by their '.Dir' slot.
}
\keyword{internal}

