\docType{methods}
\name{getHeader}
\alias{getHeader}
\alias{header}
\title{Get the header from a GenBank record.}
\usage{
getHeader(x, ...)

header(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}},
  \code{\linkS4class{gbFeature}}, or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
An \code{\linkS4class{gbHeader}} object.
}
\description{
Get the header from a GenBank record.
}

