\docType{methods}
\name{index}
\alias{index}
\title{Get indices of GenBank features}
\usage{
index(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
A numeric vector of feature indeces.
}
\description{
Get indices of GenBank features
}

