\name{isValidDb}
\alias{isValidDb}
\title{validate a gbRecord database (i.e. check if the db directory contains
all fields).}
\usage{
  isValidDb(x, verbose = TRUE, fields = GBFIELDS)
}
\description{
  validate a gbRecord database (i.e. check if the db
  directory contains all fields).
}
\keyword{internal}

