% Built by roxygen3 0.1
\name{key}
\alias{key}
\alias{key<-}
\title{Get/set feature keys from a GenBank Record}
\usage{
  key(x, attributes = FALSE, ...)

  key(x, value, ...) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
\code{\linkS4class{gbFeatureList}} instance.}

  \item{attributes}{Set the \code{accession}, \code{definition},
\code{database} attributes.}

  \item{...}{Additional arguments passed to methods.}
}

\description{
Get/set feature keys from a GenBank Record
}
\section{Method signatures for generic key}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
}

}

\section{Method signatures for generic key<-}{
\itemize{
  \item \linkS4class{gbFeature}
}

}

\keyword{methods}
