\name{listUniqueQualifs}
\alias{listUniqueQualifs}
\title{Quickly list all qualifier names}
\usage{
listUniqueQualifs(x)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}},
  \code{\linkS4class{gbFeatureList}}, or,
  \code{\linkS4class{gbFeature}} instance}
}
\value{
A character vector of qualifier names
}
\description{
Quickly list all qualifier names
}

