% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{note}
\alias{note}
\title{Return the \emph{note} qualifiers from GenBank features.}
\usage{
note(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of \emph{note}s.
}
\description{
Return the \emph{note} qualifiers from GenBank features.
}

