\name{pad}
\alias{pad}
\title{Pad a string}
\usage{
pad(x, n = 10, where = "left", pad = " ")
}
\arguments{
  \item{x}{Input character vector.}

  \item{n}{Pad \code{x} to this (minimum) width.}

  \item{where}{Side where the padding is added.}

  \item{pad}{Padding character.}
}
\value{
A character vector.
}
\description{
Pad a string
}
\keyword{internal}

