% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{product}
\alias{product}
\title{Return the \emph{product} qualifiers from GenBank features.}
\usage{
product(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of \emph{product}s.
}
\description{
Return the \emph{product} qualifiers from GenBank features.
}

