% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{proteinID}
\alias{proteinID}
\title{Return the \emph{protein_id} qualifiers from GenBank features.}
\usage{
proteinID(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of \emph{protein_id}s.
}
\description{
Return the \emph{protein_id} qualifiers from GenBank features.
}

