\docType{methods}
\name{revcompFeatures}
\alias{revcompFeatures}
\title{reverse complement features in a GenBank record}
\arguments{
  \item{x}{A gbRecord or complete gbFeatureList object
  (including the 'source' field)}

  \item{update_db}{Update filehas database with new feature
  locations.}
}
\value{
  A gbFeatureList object
}
\description{
  reverse complement features in a GenBank record
}

