% Built by roxygen3 0.1
\name{seqinfo}
\alias{accession}
\alias{definition}
\alias{seqinfo}
\alias{seqlengths}
\title{Get sequence information about genomic features}
\usage{
  seqinfo(x)

  accession(x)

  definition(x)

  seqlengths(x)
}
\description{
Get sequence information about genomic features
}
\section{Method signatures for generic seqinfo}{
\itemize{
  \item \linkS4class{GappedAlignmentPairs}
  \item \linkS4class{GappedAlignments}
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{GRanges}
  \item \linkS4class{GRangesList}
  \item \linkS4class{List}
  \item \linkS4class{RangedData}
  \item \linkS4class{RangesList}
  \item \linkS4class{SummarizedExperiment}
}

}

\section{Method signatures for generic accession}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbLocation}
  \item \linkS4class{gbRecord}
}

}

\section{Method signatures for generic definition}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
}

}

\section{Method signatures for generic seqlengths}{
\itemize{
  \item \linkS4class{ANY}
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{Seqinfo}
}

}

\keyword{methods}
