\docType{methods}
\name{getSequence}
\alias{getSequence}
\title{Get the sequence from a GenBank record.}
\usage{
getSequence(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}},
  \code{\linkS4class{gbFeature}}, or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
An \code{\linkS4class{XStringSet}} object.
}
\description{
Get the sequence from a GenBank record.
}

