\name{sequence}
\alias{sequence}
\alias{sequence-methods}
\title{Get sequences of a GenBank records or features.}
\usage{
  sequence(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}},
  \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
  An \code{\linkS4class{XStringSet}} object.
}
\description{
  Get sequences of a GenBank records or features.
}

