% Built by roxygen3 0.1
\docType{methods}
\name{start,gbFeature-method}
\alias{start,gbFeature-method}
\title{Get the start of genomic features}
\usage{
  \S4method{start}{gbFeature}(x, ...)
}
\arguments{
  \item{x}{A \code{gbFeature} object.}

  \item{join}{Join compound genomic locations into a single range.}

  \item{...}{Further arguments passed to methods.}
}

\value{
An integer vector
}
\description{
Get the start of genomic features
}
\keyword{methods}
