% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{start<-}
\alias{start<-}
\title{Set the start position of genomic features}
\usage{
start(x, ...) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}} or \code{\linkS4class{gbFeatureTable}}
object.}

\item{value}{The start information to set on \code{x}.}

\item{check}{if \code{FALSE}, don't perform validity checks.}
}
\description{
Set the start position of genomic features
}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))

cds <- x["CDS"]
start(cds) <- 10
start(cds)
}

