% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeature-class.R, R/gbFeatureTable-class.R, R/gbLocation-class.R, R/gbRecord-class.R, R/gbRecordList-class.R
\docType{methods}
\name{start}
\alias{start}
\alias{start,gbFeature-method}
\alias{start,gbFeatureTable-method}
\alias{start,gbLocation-method}
\alias{start,gbRecord-method}
\alias{start,gbRecordList-method}
\title{Get the start position of genomic features}
\usage{
start(x, ...)

\S4method{start}{gbFeature}(x, join = FALSE)

\S4method{start}{gbFeatureTable}(x, join = FALSE)

\S4method{start}{gbLocation}(x, join = FALSE)

\S4method{start}{gbRecord}(x, join = FALSE)

\S4method{start}{gbRecordList}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{join}{Join compound genomic locations into a single range.}
}
\value{
An integer vector or a list of integer vectors.
}
\description{
Get the start position of genomic features
}
\section{Methods (by class)}{
\itemize{
\item \code{gbFeature}: 

\item \code{gbFeatureTable}: 

\item \code{gbLocation}: 

\item \code{gbRecord}: 

\item \code{gbRecordList}: 
}}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))

start(x)
cds <- x["CDS"]
start(cds)
}
\seealso{
\code{\link{end}}, \code{\link{strand}}, \code{\link{span}}, \code{\link{ranges}}
}

