% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbLocation-class.R, R/gbFeature-class.R
\docType{methods}
\name{strand}
\alias{strand}
\alias{strand,gbFeature-method}
\alias{strand,gbLocation-method}
\alias{strand<-}
\alias{strand<-,gbLocation,ANY-method}
\title{Get or set the strand of genomic features}
\usage{
strand(x, ...)

strand(x, ...) <- value

\S4method{strand}{gbLocation}(x, join = FALSE)

\S4method{strand}{gbLocation,ANY}(x) <- value

\S4method{strand}{gbFeature}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{join}{Join compound genomic locations into a single range.}
}
\value{
An integer vector (or a list thereof) of 1 (plus strand), -1 (minus strand), or
\code{NA}
}
\description{
Get or set the strand of genomic features
}
\section{Methods (by class)}{
\itemize{
\item \code{gbLocation}: The strand of GenBank locations.

\item \code{x = gbLocation,value = ANY}: Set the strand of GenBank locations.

\item \code{gbFeature}: Get the strand of GenBank features.
}}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))
strand(x)
}
\seealso{
\code{\link{start}}, \code{\link{end}}, \code{\link{span}}, \code{\link{ranges}}
}

