% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{translation}
\alias{translation}
\title{Return the \emph{translation}s from GenBank features.}
\usage{
translation(x)
}
\arguments{
\item{x}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
An \code{\linkS4class{AAStringSet}}.
}
\description{
Return the \emph{translation}s from GenBank features.
}

