\name{view}
\alias{view}
\alias{view-methods}
\title{View all features in a gbFeatureList}
\usage{
  view(x, n, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeatureList}} instance.}

  \item{n}{How many features to show (Default: all).}

  \item{...}{Additional arguments passed to methods.}
}
\value{
  NULL
}
\description{
  View all features in a gbFeatureList
}

