% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{methods}
\name{view}
\alias{view}
\title{View all features in a \code{gbFeatureTable}}
\usage{
view(x, n, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeatureTable}} instance.}

  \item{n}{How many features to show (Default: all).}

  \item{...}{Additional arguments passed to methods.}
}
\description{
View all features in a \code{gbFeatureTable}
}
\keyword{internal}

