\docType{methods}
\name{width}
\alias{width}
\title{Get the width of genomic features}
\usage{
width(x, ...)
}
\arguments{
  \item{x}{A \code{gbFeature} or
  \code{gbFeatureList}object.}

  \item{...}{Further arguments passed to methods.}
}
\value{
An integer vector.
}
\description{
Get the width of genomic features
}

