% Built by roxygen3 0.1
\name{width}
\alias{width}
\title{Get the width of genomic features}
\usage{
  width(x, ...)
}
\arguments{
  \item{x}{A \code{gbFeature} or \code{gbFeatureList}object.}

  \item{...}{Further arguments passed to methods.}
}

\value{
An integer vector.
}
\description{
Get the width of genomic features
}
\section{Method signatures for generic width}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbLocation}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\keyword{methods}
