% Built by roxygen3 0.1
\name{write_genbank}
\alias{write_genbank}
\title{General function for writing out GenBank flat files}
\usage{
  write_genbank(db, outfile = "out.gbk")
}
\arguments{
  \item{db}{A \code{\linkS4class{gbRecord}} object.}

  \item{outfile}{Output file.}
}

\description{
General function for writing out GenBank flat files
}
\details{
For a description of the GenBank format see
\url{http://www.ncbi.nlm.nih.gov/collab/FT/}
}
