\name{writeGB}
\alias{writeGB}
\title{General function for writing out GenBank flat files}
\usage{
  writeGB(db, outfile = "out.gbk")
}
\arguments{
  \item{db}{A \code{gbRecord} object.}

  \item{outfile}{Output file.}
}
\description{
  For a description of the GenBank format see
  \url{http://www.ncbi.nlm.nih.gov/collab/FT/}
}

