% Built by roxygen3 0.1
\name{getLocus}
\alias{getAccession}
\alias{getComment}
\alias{getDBLink}
\alias{getDBSource}
\alias{getDate}
\alias{getDefinition}
\alias{getDivision}
\alias{getGeneID}
\alias{getKeywords}
\alias{getLength}
\alias{getLocus}
\alias{getMoltype}
\alias{getOrganism}
\alias{getReference}
\alias{getSource}
\alias{getTaxonomy}
\alias{getTopology}
\alias{getVersion}
\title{Access slots of aGenBank record object.}
\usage{
  getLocus(x)

  getLength(x)

  getMoltype(x)

  getTopology(x)

  getDivision(x)

  getDate(x)

  getDefinition(x)

  getAccession(x)

  getVersion(x)

  getGeneID(x)

  getDBLink(x)

  getDBSource(x)

  getSource(x)

  getOrganism(x)

  getTaxonomy(x)

  getKeywords(x)

  getReference(x)

  getComment(x)
}
\description{
Access slots of aGenBank record object.
}
\section{Method signatures for generic getLocus}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getLength}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getMoltype}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getTopology}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getDivision}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getDate}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getDefinition}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getAccession}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbLocation}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getVersion}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getGeneID}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getDBLink}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getDBSource}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getSource}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getOrganism}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getTaxonomy}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getKeywords}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getReference}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\section{Method signatures for generic getComment}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
  \item \linkS4class{seqinfo}
}

}

\keyword{methods}
