% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{dbxref}
\alias{dbxref}
\alias{dbxref,gbFeature-method}
\alias{dbxref,gbFeatureTable-method}
\alias{dbxref,gbRecord-method}
\title{Access the \code{db_xref}s of GenBank features}
\usage{
dbxref(x, db = NULL, ...)

\S4method{dbxref}{gbFeature}(x, db = NULL, ...)

\S4method{dbxref}{gbFeatureTable}(x, db = NULL, ...)

\S4method{dbxref}{gbRecord}(x, db = NULL, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}}, or
\code{\linkS4class{gbRecord}} object.}

\item{db}{(Optional) A character vector giving the database names of the
desired \code{db_xref}s.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named character vector (or list of named character vectors)
of db_xrefs.
}
\description{
Access the \code{db_xref}s of GenBank features
}
\examples{
gbk_file <- system.file("extdata", "marine_metagenome.gbk", package = "biofiles")
x <- gbRecord(gbk_file)

## all db_xrefs associated with CDSs
dbxref(x["CDS"])

## retrieve the TaxId from the "source" field.
dbxref(x[[1]], "taxon")
}

