\docType{methods}
\name{dbxref}
\alias{dbxref}
\title{Get the \code{db_xref}s of GenBank features}
\usage{
dbxref(x, db = NULL, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{db}{(Optional) A character vector giving the
  database names of the desired \code{db_xref}s.}

  \item{...}{Additional arguments passed to methods.}
}
\description{
Get the \code{db_xref}s of GenBank features
}

