\name{dbxref}
\alias{dbxref}
\alias{dbxref-methods}
\title{Get/set db_xref from a GenBank record}
\usage{
  dbxref(x, db = NULL, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} object.}

  \item{db}{(Optional) Character vector giving the database
  names of the desired db_xrefs.}

  \item{...}{Additional arguments passed to methods.}
}
\description{
  Get/set db_xref from a GenBank record
}

