% Built by roxygen3 0.1
\name{dbxref}
\alias{dbxref}
\title{Get/set db_xref from a GenBank record}
\usage{
  dbxref(x, db = NULL, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
\code{\linkS4class{gbFeatureList}} instance.}

  \item{db}{(Optional) A character vector giving the database names of the
desired db_xrefs.}

  \item{...}{Additional arguments passed to methods.}
}

\description{
Get/set db_xref from a GenBank record
}
\section{Method signatures for generic dbxref}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
}

}

\keyword{methods}
