% Built by roxygen3 0.1
\name{features}
\alias{features}
\title{Retrieve features of a GenBank record.}
\usage{
  features(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}} instance.}

  \item{...}{Additional arguments passed to methods.}
}

\value{
The \code{\linkS4class{gbFeatureList}} object contained in a
gbRecord database.
}
\description{
Retrieve features of a GenBank record.
}
\section{Method signatures for generic features}{
\itemize{
  \item \linkS4class{gbRecord}
}

}

\keyword{methods}
