\docType{methods}
\name{fuzzy}
\alias{fuzzy}
\alias{fuzzy,gbFeature-method}
\alias{fuzzy,gbFeatureList-method}
\alias{fuzzy,gbRecord-method}
\alias{fuzzy,gbRecordList-method}
\title{Has a feature unclear start/end positions?}
\usage{
fuzzy(x)

\S4method{fuzzy}{gbFeature}(x)

\S4method{fuzzy}{gbFeatureList}(x)

\S4method{fuzzy}{gbRecord}(x)

\S4method{fuzzy}{gbRecordList}(x)
}
\description{
Has a feature unclear start/end positions?
}

