% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbLocation-class.R, R/gbFeature-class.R
\docType{methods}
\name{fuzzy}
\alias{fuzzy}
\alias{fuzzy,gbFeature-method}
\alias{fuzzy,gbLocation-method}
\title{Has a feature fuzzy locations?}
\usage{
fuzzy(x, ...)

\S4method{fuzzy}{gbLocation}(x)

\S4method{fuzzy}{gbFeature}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A logical matrix.
}
\description{
With a GenBank location like \emph{complement(<123..150)} we don't know the exact
start position of the feature. Use \code{fuzzy} to test for fuzzy locations.
}
\section{Methods (by class)}{
\itemize{
\item \code{gbLocation}: Has a  GenBank location a fuzzy start or end?

\item \code{gbFeature}: Has a  GenBank feature a fuzzy start or end?
}}
\examples{
l <- as.gbLocation("complement(<123..150)")
fuzzy(l)

## note that start() or end() return exact positions even if they are fuzzy.
start(l)
}

