% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{gbFeatureTable-class}
\alias{gbFeatureTable-class}
\title{Class \code{"gbFeatureTable"}}
\description{
\dQuote{gbFeatureTable} is an S4 class that provides a container for
\dQuote{\linkS4class{gbFeature}}s retrived from GenBank flat files.
}
\section{Slots}{

\describe{
\item{\code{.seqinfo}}{A \code{\linkS4class{seqinfo}} object containing the
genome sequence as an \code{\linkS4class{XStringSet}} object and
sequence metadata as a \code{\linkS4class{gbHeader}} object.}

\item{\code{.id}}{An integer vector the indices of the \code{\linkS4class{gbFeature}}s
contained within a \code{gbFeatureTable} object.}

\item{\code{.Data}}{A list of \code{\linkS4class{gbFeature}} objects.}
}}

