% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{class}
\name{gbRecord-class}
\alias{gbRecord-class}
\title{Class \code{"gbRecord"}}
\description{
\dQuote{gbRecord} is an S4 class that provides a container
for data parsed from a GenBank or GenPept records. For
instantiation of a gbRecord object use the import function
\code{\link{gbRecord}}.
}
\section{Slots}{

\describe{
\item{\code{seqinfo}}{A \code{"\linkS4class{seqinfo}"} instance; This is a
reference class holding the sequence as an \code{"\linkS4class{XStringSet}"}
instance and header of the file containing metadata as a
\code{"\linkS4class{gbHeader}"} object.}

\item{\code{features}}{A \code{"\linkS4class{gbFeatureTable}"} instance.}

\item{\code{contig}}{If present, a CONTIG record.}
}}
\section{Accessor functions}{
  \code{\link{header}}, \code{\link{ft}},
  \code{\link{ranges}}
}
\seealso{
The constructor, \code{\link{gbRecord}}
}

