% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gbRecord-class.R
\docType{class}
\name{gbRecord-class}
\alias{gbRecord-class}
\alias{new_gbRecord}
\title{Class \code{"gbRecord"}}
\description{
\dQuote{gbRecord} is an S4 class that provides a container for data
parsed from GenBank, GenPept, Embl or IMGT records. For instantiation of a
gbRecord object use the import function \code{\link{gbRecord}}.
}
\section{Slots}{

\describe{
\item{\code{seqinfo}}{A \code{"\linkS4class{seqinfo}"} instance; This is a
reference class holding the sequence as an \code{"\linkS4class{XStringSet}"}
instance and header of the file containing metadata as a
\code{"\linkS4class{gbHeader}"} object.}

\item{\code{features}}{A \code{"\linkS4class{gbFeatureTable}"} instance.}

\item{\code{contig}}{If present, a CONTIG record.}
}}
\section{Accessor functions}{

 \code{\link{header}}, \code{\link{ft}}, \code{\link{ranges}}
}
\seealso{
The constructor, \code{\link{gbRecord}}
}

