% Built by roxygen3 0.1
\docType{class}
\name{gbRecordList-class}
\alias{gbRecordList}
\alias{gbRecordList-class}
\title{gbRecordList}
\usage{
  gbRecordList(...)
}
\arguments{
  \item{...}{\dQuote{\linkS4class{gbRecord}} elements.}
}

\description{
\dQuote{gbRecordList} is an S4 class that provides a container for
\dQuote{\linkS4class{gbRecord}}s retrived from GenBank flat files.
}
\section{Class hierarchy for gbRecordList}{
Super classes:
\itemize{
  \item \linkS4class{list}
  \item \linkS4class{vector}
  \item \linkS4class{vectorORfactor}
}

}

\section{Generics with methods for gbRecordList}{
\itemize{
  \item \code{\link{accession}}
  \item \code{\link{definition}}
  \item \code{\link{end}}
  \item \code{\link{features}}
  \item \code{\link{ranges}}
  \item \code{\link{seqinfo}}
  \item \code{\link{seqlengths}}
  \item \code{\link{sequence}}
  \item \code{\link{show}}
  \item \code{\link{start}}
  \item \code{\link{strand}}
  \item \code{\link{summary}}
  \item \code{\link{width}}
}

}

\keyword{classes}
