\name{.gbReferenceList}
\alias{.gbReferenceList}
\title{Generator object for the \code{"\linkS4class{gbReferenceList}"} class}
\usage{
.gbReferenceList(...)
}
\arguments{
  \item{...}{List of arguments}
}
\description{
The generator object for the
\code{"\linkS4class{gbReferenceList}"} reference class.
}
\section{Methods}{
  \describe{ \item{\code{#new(ref)}:}{ Create a new
  \code{\linkS4class{gbReferenceList}} object }
  \item{\code{#to_string(write_to_file = FALSE)}:}{ Create
  a string representation of a GenBank REFERENCE list. } }
}
\seealso{
\code{"\linkS4class{gbReferenceList}"}
}
\keyword{classes}
\keyword{internal}

