% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getContigSeq}
\alias{getContigSeq}
\title{Retrieve the sequence of a contig}
\usage{
getContigSeq(x, merge = TRUE)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}} object.}

  \item{merge}{Merge the retrieved contig sequences.}
}
\value{
A \code{\linkS4class{DNAStringSet}} instance.
}
\description{
## EXPERIMENTAL ##
}
\details{
For GenBank records that contain a CONTIG field, try to get the contig
sequences and (optionally) merge them into a single sequence.
}
\examples{
###
}

