\docType{methods}
\name{getFeatures}
\alias{getFeatures}
\title{Retrieve features of a GenBank record.}
\arguments{
  \item{data}{An instance of \code{\link{gbRecord-class}}.}
}
\value{
  The \code{\link{gbFeatureList-class}} object contained in
  a gbRecord database.
}
\description{
  Retrieve features of a GenBank record.
}
\seealso{
  Other accessors: \code{\link{getIndex}},
  \code{\link{getLocation}}
}

