% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeature-class.R,
%   R/gbFeatureTable-class.R, R/gbRecord-class.R
\name{getHeader}
\alias{getHeader}
\alias{header}
\alias{header,gbFeature-method}
\alias{getHeader,gbFeature-method}
\alias{getHeader,gbFeatureTable-method}
\alias{header,gbFeatureTable-method}
\alias{getHeader,gbRecord-method}
\alias{header,gbRecord-method}
\title{Extract the header from a \code{"\linkS4class{gbRecord}"} object.}
\usage{
getHeader(x, ...)

header(x, ...)

\S4method{header}{gbFeature}(x)

\S4method{getHeader}{gbFeature}(x)

\S4method{getHeader}{gbFeatureTable}(x)

\S4method{header}{gbFeatureTable}(x)

\S4method{getHeader}{gbRecord}(x)

\S4method{header}{gbRecord}(x)
}
\arguments{
\item{x}{A \code{"\linkS4class{gbRecord}"}, \code{"\linkS4class{gbFeature}"},
or \code{"\linkS4class{gbFeatureTable}"} instance.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{"\linkS4class{gbHeader}"} instance
}
\description{
Extract the header from a \code{"\linkS4class{gbRecord}"} object.
}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))
getHeader(x)
header(x)
}
