\name{getKey}
\alias{getKey}
\alias{getKey-methods}
\title{Retrieve feature keys from a GenBank record}
\usage{
  getKey(x, attributes = FALSE)
}
\arguments{
  \item{x}{A \code{gbFeature} or \code{gbFeatureList}
  object.}

  \item{attributes}{Set the \code{accession},
  \code{definition}, \code{database} attributes.}
}
\description{
  Retrieve feature keys from a GenBank record
}

