\docType{methods}
\name{getLocation}
\alias{getLocation}
\title{get genomic location of a GenBank feature}
\arguments{
  \item{x}{A gbFeature or gbFeatureList object}
}
\value{
  A data frame
}
\description{
  get genomic location of a GenBank feature
}
\seealso{
  Other accessors: \code{\link{getFeatures}},
  \code{\link{getIndex}}
}

