\docType{methods}
\name{getLocation}
\alias{getLocation}
\title{get genomic location of a GenBank feature}
\arguments{
  \item{x}{A \code{\link{gbFeature}} or
  \code{\link{gbFeatureList}} object}

  \item{attributes}{set the \code{accession},
  \code{definition}, \code{database} attributes.}

  \item{join}{combine compound locations}
}
\value{
  A data frame
}
\description{
  get genomic location of a GenBank feature
}

