\docType{methods}
\name{getQualifier}
\alias{getQualifier}
\title{Retrieve feature keys from a GenBank record}
\arguments{
  \item{x}{A \code{gbFeature} or \code{gbFeatureList}
  object.}

  \item{which}{Character vector giving the name(s) of the
  qualifiers to retrieve.}

  \item{attributes}{Set the \code{accession},
  \code{definition}, \code{database} attributes.}
}
\description{
  Retrieve feature keys from a GenBank record
}

