% Built by roxygen3 0.1
\name{hasValidDb}
\alias{hasValidDb}
\title{validate a gbRecord database which if referenced by gbFeature or
gbFeatureList objects by their '.Dir' slot.}
\usage{
  hasValidDb(object, verbose = TRUE)
}
\description{
validate a gbRecord database which if referenced by gbFeature or
gbFeatureList objects by their '.Dir' slot.
}
\keyword{internal}
