% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeature-class.R,
%   R/gbFeatureTable-class.R, R/gbRecord-class.R, R/gbRecordList-class.R
\name{key}
\alias{key}
\alias{key<-}
\alias{key,gbFeature-method}
\alias{key<-,gbFeature-method}
\alias{key,gbFeatureTable-method}
\alias{key,gbRecord-method}
\alias{key,gbRecordList-method}
\title{Get/set keys of GenBank features}
\usage{
key(x, ...)

key(x, check = TRUE) <- value

\S4method{key}{gbFeature}(x)

\S4method{key}{gbFeature}(x, check = TRUE) <- value

\S4method{key}{gbFeatureTable}(x)

\S4method{key}{gbRecord}(x)

\S4method{key}{gbRecordList}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Additional arguments passed to methods.}

\item{check}{if \code{FALSE}, don't perform validity checks.}

\item{value}{The key information to set on \code{x}.}
}
\description{
Get/set keys of GenBank features
}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))
key(x)
}
