% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{methods}
\name{key}
\alias{key}
\alias{key,gbFeature-method}
\alias{key,gbFeatureTable-method}
\alias{key,gbRecord-method}
\alias{key,gbRecordList-method}
\alias{key<-}
\alias{key<-,gbFeature-method}
\title{Get/set keys of GenBank features}
\usage{
key(x, ...)

key(x, check = TRUE) <- value

\S4method{key}{gbFeature}(x)

\S4method{key}{gbFeatureTable}(x)

\S4method{key}{gbRecord}(x)

\S4method{key}{gbRecordList}(x)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}},
  \code{\linkS4class{gbFeatureTable}},
  \code{\linkS4class{gbRecord}}, or
  \code{\linkS4class{gbRecordList}} object.}

  \item{...}{Additional arguments passed to methods.}

  \item{check}{if \code{FALSE}, don't perform validity
  checks.}

  \item{value}{The key information to set on \code{x}.}
}
\description{
Get/set keys of GenBank features
}
\examples{
gbk_file <- system.file("extdata", "marine_metagenome.gbk", package = "biofiles")
x <- gbRecord(gbk_file)
key(x)
}

