% Built by roxygen3 0.1
\name{key}
\alias{key}
\alias{key<-}
\title{Get/set keys of GenBank features}
\usage{
  key(x, ...)

  key(x, value, ...) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
\code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}

\description{
Get/set keys of GenBank features
}
\section{Method signatures for generic key}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
}

}

\section{Method signatures for generic key<-}{
\itemize{
  \item \linkS4class{gbFeature}
}

}

\keyword{methods}
