\docType{methods}
\name{key}
\alias{key}
\alias{key<-}
\title{Get/set keys of GenBank features}
\usage{
key(x, ...)

key(x, ...) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\description{
Get/set keys of GenBank features
}

