\name{key}
\alias{key}
\alias{key-methods}
\title{Get/set feature keys from a GenBank Record}
\usage{
  key(x, attributes = FALSE)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} object.}

  \item{attributes}{Set the \code{accession},
  \code{definition}, \code{database} attributes.}

  \item{...}{Additional arguments passed to methods.}
}
\description{
  Get/set feature keys from a GenBank Record
}

