% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbRecord-class.R
\docType{methods}
\name{length,gbRecord-method}
\alias{length,gbRecord-method}
\title{Get the number of gbFeatures.}
\usage{
\S4method{length}{gbRecord}(x)
}
\arguments{
\item{x}{A \code{gbRecord}}
}
\value{
An integer
}
\description{
Get the number of gbFeatures.
}

