\docType{methods}
\name{location}
\alias{location}
\title{Get genomic locations of GenBank features}
\usage{
location(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}
\value{
A list of \code{\linkS4class{gbLocation}} objects
}
\description{
Get genomic locations of GenBank features
}

