% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{locusTag}
\alias{locusTag}
\title{Return the \emph{locus_tag} qualifiers from GenBank features.}
\usage{
locusTag(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of \emph{locus_tag}s.
}
\description{
Return the \emph{locus_tag} qualifiers from GenBank features.
}

