% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{qualif}
\alias{geneID}
\alias{locusTag}
\alias{note}
\alias{product}
\alias{proteinID}
\alias{qualif}
\alias{qualif,gbFeature-method}
\alias{qualif,gbFeatureTable-method}
\alias{qualif,gbRecord-method}
\alias{qualif<-}
\alias{qualif<-,gbFeature-method}
\alias{translation}
\title{Get/set qualifiers of GenBank features}
\usage{
qualif(x, which = "", ...)

qualif(x, which, check = TRUE) <- value

\S4method{qualif}{gbFeature}(x, which, fixed = FALSE, use.names = TRUE)

\S4method{qualif}{gbFeatureTable}(x, which = "", fixed = FALSE,
  use.names = TRUE)

\S4method{qualif}{gbRecord}(x, which = "", fixed = FALSE,
  use.names = TRUE)

locusTag(x)

product(x)

note(x)

proteinID(x)

geneID(x)

translation(x)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}}, or
\code{\linkS4class{gbRecord}} object.}

\item{which}{A character vector giving the name(s) of the
qualifiers to retrieve or set.}

\item{fixed}{If \code{TRUE}, \code{which} is matched against qualifiers as is,
if \code{FALSE} it is treated as a regular expression.}

\item{use.names}{If \code{TRUE}, return a \code{data.frame} using \code{which}
as column names, if \code{FALSE} return, if possible, a character vector or
a list.}

\item{...}{Additional arguments passed to methods.}

\item{check}{if \code{FALSE}, don't perform validity checks.}

\item{value}{The qualifier information to set on \code{x}.}
}
\value{
A \code{data.frame}.
}
\description{
Get/set qualifiers of GenBank features
}
\examples{
gbk_file <- system.file("extdata", "marine_metagenome.gbk", package = "biofiles")
x <- gbRecord(gbk_file)
qualif(x[[1]], 'db_xref')

## use shortcuts to common qualifiers
proteinID(x["CDS"])
}

