\name{qualif}
\alias{qualif}
\alias{qualif-methods}
\title{Get/set feature qualifiers from a GenBank record}
\usage{
  qualif(x, which, attributes = FALSE)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeature}} or
  \code{\linkS4class{gbFeatureList}} object.}

  \item{which}{(Optional) Character vector giving the
  name(s) of the qualifiers to retrieve.}

  \item{attributes}{Set the \code{accession},
  \code{definition}, \code{database} attributes.}

  \item{...}{Additional arguments passed to methods.}
}
\description{
  Get/set feature qualifiers from a GenBank record
}

