% Generated by roxygen2 (3.1.0.99): do not edit by hand
\docType{methods}
\name{revcomp}
\alias{revcomp}
\alias{revcomp,gbFeatureTable-method}
\alias{revcomp,gbRecord-method}
\title{Reverse-complement features in a GenBank record}
\usage{
revcomp(x, order = TRUE)

\S4method{revcomp}{gbFeatureTable}(x, order = TRUE)

\S4method{revcomp}{gbRecord}(x, order = TRUE)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeatureTable}} or
  \code{\linkS4class{gbRecord}} object (gbFeatureTables
  must be complete and include a 'source' field).}

  \item{order}{Reorder features after reverse-complementing
  them.}
}
\description{
Reverse-complement features in a GenBank record
}
\examples{
gbk_file <- system.file("extdata", "S_cerevisiae_mito.gbk", package="biofiles")
x <- gbRecord(gbk_file)
xr <- revcomp(x)
}

