\docType{methods}
\name{select}
\alias{select}
\title{Select elements from a GenBank Record}
\arguments{
  \item{x}{A \sQuote{\code{gbRecord}} or
  \sQuote{\code{gbFeatureList}} object}

  \item{subset}{Which elements to select from indicated by
  index, key, location, or qualifier value.}

  \item{select}{Which information to be retrieved from the
  subsetted elements.}
}
\description{
  Select elements from a GenBank Record
}

