% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbHeader-class.R
\docType{class}
\name{seqinfo-class}
\alias{seqinfo-class}
\alias{seqinfo}
\alias{seqinfo-class}
\title{Generator object for the \code{"seqinfo"} reference class}
\arguments{
\item{...}{List of arguments; must be named arguments
corresponding to the fields of a \code{\linkS4class{gbHeader}} object}
}
\description{
Generator object for the \code{"seqinfo"} reference class

Generates a container for header and sequence information.
}

\section{Methods}{

\describe{
\item{\code{#new(header = NULL, sequence = NULL)}:}{
   Create a new \code{\linkS4class{gbHeader}} object}
}
}

\section{Fields}{

\describe{
\item{\code{header}:}{ A \code{\linkS4class{gbHeader}} object or \code{NULL}. }
\item{\code{sequence}:}{  A \code{\linkS4class{XStringSet}} object or \code{NULL}. }
}
}

\section{Extends}{
 All reference classes extend and inherit methods from
   \code{"\linkS4class{envRefClass}"}.
}

\examples{
showClass("seqinfo")
}
\keyword{classes}
\keyword{internal}
