% Built by roxygen3 0.1
\docType{class}
\name{seqinfo-class}
\alias{seqinfo-class}
\title{seqinfo-class}
\description{
seqinfo-class
}
\section{Class hierarchy for seqinfo}{
Super classes:
\itemize{
  \item \linkS4class{.environment}
  \item \linkS4class{environment}
  \item \linkS4class{refObject}
}

}

\section{Generics with methods for seqinfo}{
\itemize{
  \item \code{\link{.header}}
  \item \code{\link{.sequence}}
  \item \code{\link{getAccession}}
  \item \code{\link{getComment}}
  \item \code{\link{getDate}}
  \item \code{\link{getDBLink}}
  \item \code{\link{getDBSource}}
  \item \code{\link{getDefinition}}
  \item \code{\link{getDivision}}
  \item \code{\link{getGeneID}}
  \item \code{\link{getKeywords}}
  \item \code{\link{getLength}}
  \item \code{\link{getLocus}}
  \item \code{\link{getMoltype}}
  \item \code{\link{getOrganism}}
  \item \code{\link{getReference}}
  \item \code{\link{getSequence}}
  \item \code{\link{getSource}}
  \item \code{\link{getTaxonomy}}
  \item \code{\link{getTopology}}
  \item \code{\link{getVersion}}
  \item \code{\link{initialize}}
  \item \code{\link{show}}
}

}

\keyword{classes}
