% Built by roxygen3 0.1
\name{seqinfo}
\alias{seqinfo}
\title{Get sequence information about genomic features}
\usage{
  seqinfo(x)
}
\description{
Get sequence information about genomic features
}
\section{Method signatures for generic seqinfo}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\keyword{methods}
