\name{seqinfo}
\alias{seqinfo}
\title{Generator object for the \code{\linkS4class{seqinfo}} class}
\usage{
seqinfo(...)
}
\arguments{
  \item{...}{List of arguments; must be named arguments
  corresponding to the fields of a
  \code{\linkS4class{gbHeader}} object}
}
\description{
The generator object for the \code{\linkS4class{seqinfo}}
reference class.
}
\section{Methods}{
  \describe{ \item{\code{#new(header = NULL, sequence =
  NULL)}:}{ Create a new \code{\linkS4class{gbHeader}}
  object} }
}
\seealso{
\code{\linkS4class{seqinfo}}
}
\keyword{classes}
\keyword{internal}

