% Built by roxygen3 0.1
\name{seqinfo}
\alias{accession}
\alias{definition}
\alias{geneid}
\alias{seqinfo}
\alias{seqlengths}
\title{Get sequence information about genomic features}
\usage{
  seqinfo(x)

  accession(x)

  geneid(x, db = 'gi')

  definition(x)

  seqlengths(x)
}
\description{
Get sequence information about genomic features
}
\section{Method signatures for generic seqinfo}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\section{Method signatures for generic accession}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbLocation}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\section{Method signatures for generic geneid}{
\itemize{
  \item \linkS4class{gbRecord}
}

}

\section{Method signatures for generic definition}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\section{Method signatures for generic seqlengths}{
\itemize{
  \item \linkS4class{ANY}
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
  \item \linkS4class{gbRecordList}
}

}

\keyword{methods}
