% Built by roxygen3 0.1
\name{sequence}
\alias{sequence}
\title{Get sequences of GenBank features}
\usage{
  sequence(x, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeature}},
 or \code{\linkS4class{gbFeatureList}} instance.}

  \item{...}{Additional arguments passed to methods.}
}

\value{
An \code{\linkS4class{XStringSet}} object.
}
\description{
Get sequences of GenBank features
}
\section{Method signatures for generic sequence}{
\itemize{
  \item \linkS4class{gbFeature}
  \item \linkS4class{gbFeatureList}
  \item \linkS4class{gbRecord}
}

}

\keyword{methods}
