% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{start}
\alias{start}
\alias{start,gbFeature-method}
\alias{start,gbFeatureTable-method}
\alias{start,gbLocation-method}
\alias{start,gbRecord-method}
\alias{start,gbRecordList-method}
\alias{start<-}
\alias{start<-,gbFeature-method}
\alias{start<-,gbFeatureTable-method}
\alias{start<-,gbLocation-method}
\title{Get or set the start position of genomic features}
\usage{
start(x, ...)

start(x, check = TRUE) <- value

\S4method{start}{gbFeature}(x, join = FALSE)

\S4method{start}{gbFeature}(x, check = TRUE) <- value

\S4method{start}{gbFeatureTable}(x, join = FALSE)

\S4method{start}{gbFeatureTable}(x, check = TRUE) <- value

\S4method{start}{gbLocation}(x, join = FALSE)

\S4method{start}{gbLocation}(x, check = TRUE) <- value

\S4method{start}{gbRecord}(x, join = FALSE)

\S4method{start}{gbRecordList}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{check}{if \code{FALSE}, don't perform validity checks.}

\item{value}{The start information to set on \code{x}.}

\item{join}{Join compound genomic locations into a single range.}
}
\value{
An integer vector or a list of integer vectors.
}
\description{
Get or set the start position of genomic features
}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))

start(x)
cds <- x["CDS"]
start(cds)
}
\seealso{
\code{\link{end}}, \code{\link{strand}}, \code{\link{span}}, \code{\link{ranges}}
}

