% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbLocation-class.R, R/gbFeature-class.R
\docType{methods}
\name{start}
\alias{start}
\alias{start,gbFeature-method}
\alias{start,gbLocation-method}
\title{Get the start position of genomic features}
\usage{
start(x, ...)

\S4method{start}{gbLocation}(x, join = FALSE)

\S4method{start}{gbFeature}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{join}{Join compound genomic locations into a single range.}
}
\value{
An integer vector or a list of integer vectors.
}
\description{
Get the start position of genomic features
}
\section{Methods (by class)}{
\itemize{
\item \code{gbLocation}: Get the start values of GenBank locations.

\item \code{gbFeature}: Get the start values of GenBank features.
}}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))

start(x)
cds <- x["CDS"]
start(cds)
}
\seealso{
\code{\link{end}}, \code{\link{strand}}, \code{\link{span}}, \code{\link{ranges}}
}

