% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeature-class.R, R/gbFeatureTable-class.R, R/gbLocation-class.R, R/gbRecord-class.R, R/gbRecordList-class.R
\docType{methods}
\name{strand}
\alias{strand}
\alias{strand,gbFeature-method}
\alias{strand,gbFeatureTable-method}
\alias{strand,gbLocation-method}
\alias{strand,gbRecord-method}
\alias{strand,gbRecordList-method}
\alias{strand<-}
\alias{strand<-,gbFeature,ANY-method}
\alias{strand<-,gbFeatureTable,ANY-method}
\alias{strand<-,gbLocation,ANY-method}
\title{Get or set the strand of genomic features}
\usage{
strand(x, ...)

strand(x, ...) <- value

\S4method{strand}{gbFeature}(x, join = FALSE)

\S4method{strand}{gbFeature,ANY}(x) <- value

\S4method{strand}{gbFeatureTable}(x, join = FALSE)

\S4method{strand}{gbFeatureTable,ANY}(x) <- value

\S4method{strand}{gbLocation}(x, join = FALSE)

\S4method{strand}{gbLocation,ANY}(x) <- value

\S4method{strand}{gbRecord}(x, join = FALSE)

\S4method{strand}{gbRecordList}(x, join = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeature}}, \code{\linkS4class{gbFeatureTable}},
\code{\linkS4class{gbRecord}}, or \code{\linkS4class{gbRecordList}} object.}

\item{...}{Further arguments passed to methods.}

\item{join}{Join compound genomic locations into a single range.}
}
\value{
An integer vector (or a list thereof) of 1 (plus strand), -1 (minus strand), or
\code{NA}
}
\description{
Get or set the strand of genomic features
}
\section{Methods (by class)}{
\itemize{
\item \code{gbFeature}: 

\item \code{x = gbFeature,value = ANY}: 

\item \code{gbFeatureTable}: 

\item \code{x = gbFeatureTable,value = ANY}: 

\item \code{gbLocation}: 

\item \code{x = gbLocation,value = ANY}: 

\item \code{gbRecord}: 

\item \code{gbRecordList}: 
}}
\examples{
load(system.file("extdata", "marine_metagenome.rda", package = "biofiles"))
strand(x)
}
\seealso{
\code{\link{start}}, \code{\link{end}}, \code{\link{span}}, \code{\link{ranges}}
}

