% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,gbFeature-method}
\alias{summary,gbFeatureTable-method}
\alias{summary,gbRecord-method}
\alias{summary,gbRecordList-method}
\title{Summarise a GenBank record.}
\usage{
summary(object, n = 8, ...)

\S4method{summary}{gbFeature}(object, ...)

\S4method{summary}{gbFeatureTable}(object, n = 8, ...)

\S4method{summary}{gbRecord}(object, n = 7, ...)

\S4method{summary}{gbRecordList}(object, n = 2, ...)
}
\arguments{
  \item{object}{An object of
  class\code{\linkS4class{gbFeature}},
  \code{\linkS4class{gbFeatureTable}},
  \code{\linkS4class{gbRecord}}, or
  \code{\linkS4class{gbRecordList}}.}

  \item{n}{For \code{list}-like objects, how many elements
  should be summarized in head and tail.}

  \item{...}{Arguments to be passed to or from other
  methods.}
}
\description{
Summarise a GenBank record.
}
\examples{
gbk_file <- system.file("extdata", "marine_metagenome.gbk", package = "biofiles")
x <- gbRecord(gbk_file)
summary(x)
}

