% Generated by roxygen2 (4.0.0): do not edit by hand
\name{uniqueQualifs}
\alias{uniqueQualifs}
\title{Quickly list all qualifier names}
\usage{
uniqueQualifs(x)
}
\arguments{
\item{x}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance}
}
\value{
A character vector of qualifier names
}
\description{
Quickly list all qualifier names
}

