% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{uniqueQualifs}
\alias{uniqueQualifs}
\title{Quickly list all qualifier names.}
\usage{
uniqueQualifs(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of qualifier names.
}
\description{
Quickly list all qualifier names.
}

