% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{uniqueQualifs}
\alias{uniqueQualifs}
\title{Quickly list all qualifier names.}
\usage{
uniqueQualifs(...)
}
\arguments{
\item{...}{A \code{\linkS4class{gbRecord}}, \code{\linkS4class{gbFeatureTable}},
or, \code{\linkS4class{gbFeature}} instance.}
}
\value{
A character vector of qualifier names.
}
\description{
Quickly list all qualifier names.
}

