% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/gbFeatureTable-class.R
\docType{methods}
\name{view}
\alias{view}
\alias{view,gbFeatureTable-method}
\title{View all features in a \code{gbFeatureTable}}
\usage{
view(x, n, ...)

\S4method{view}{gbFeatureTable}(x, n)
}
\arguments{
\item{x}{A \code{\linkS4class{gbFeatureTable}} instance.}

\item{n}{How many features to show (Default: all).}

\item{...}{Additional arguments passed to methods.}
}
\description{
View all features in a \code{gbFeatureTable}
}
\keyword{internal}

