\docType{methods}
\name{view}
\alias{view}
\title{View all features in a \code{gbFeatureList}}
\usage{
view(x, n, ...)
}
\arguments{
  \item{x}{A \code{\linkS4class{gbFeatureList}} instance.}

  \item{n}{How many features to show (Default: all).}

  \item{...}{Additional arguments passed to methods.}
}
\description{
View all features in a \code{gbFeatureList}
}

